+++ pkg/messages/proposeTreatyMessage.go
@@ -19,15 +19,19 @@ func (msg *ProposeTreatyMessage) Treaty() Treaty {
 	return msg.treaty
 }
 
+func (msg *ProposeTreatyMessage) TreatyID() uuid.UUID {
+	return msg.treaty.ID()
+}
+
+func (msg *ProposeTreatyMessage) Reply(senderID uuid.UUID, senderFloor int, targetFloor int, response bool) TreatyResponseMessage {
+	reply := *NewTreatyResponseMessage(senderID, senderFloor, targetFloor, response, msg.treaty.ID(), msg.ID())
+	return reply
+}
+
 func (msg *ProposeTreatyMessage) Visit(a Agent) {
 	if msg.TargetFloor() != a.Floor() {
 		a.HandlePropogate(msg)
 	} else {
 		a.HandleProposeTreaty(*msg)
 	}
-}
-
-func (msg *ProposeTreatyMessage) Reply(senderID uuid.UUID, senderFloor int, targetFloor int, response bool) TreatyResponseMessage {
-	reply := *NewTreatyResponseMessage(senderID, senderFloor, targetFloor, response, msg.treaty.Id(), msg.ID())
-	return reply
 }
